Added blueprint layout
