+++ flask-api/.env.example
@@ -1,0 +1,4 @@
+## COPY CONTENTS INTO YOUR OWN .ENV FILE AND FILL IN ACCORDINGLY
+
+# Not so secret for development XD
+SECRET_KEY=d941c39d08ce8fb6d6a24bbedce0564529f45f83afeab8fc6a1b1141a26705fc
