+++ flask-api/Dockerfile
@@ -12,4 +12,4 @@ RUN pip install -r requirements.txt
 COPY . .
 
 EXPOSE 5000
-CMD [ "python", "app.py"]
+CMD [ "python", "run.py"]
