+++ flask-api/run.py
@@ -1,0 +1,5 @@
+from server import app
+
+# main driver function for dev mode
+if __name__ == "__main__":
+    app.run(debug=True, host="0.0.0.0")
